Refactoring ovdc related operations out of broker manager into ovdc manager
