+++ container_service_extension/broker_manager.py
@@ -2,8 +2,6 @@
 # Copyright (c) 2019 VMware, Inc. All Rights Reserved.
 # SPDX-License-Identifier: BSD-2-Clause
 
-from collections import namedtuple
-
 from pyvcloud.vcd.org import Org
 import requests
 
@@ -13,7 +11,6 @@ from container_service_extension.exceptions import CseDuplicateClusterError
 from container_service_extension.exceptions import CseServerError
 from container_service_extension.exceptions import PksDuplicateClusterError
 from container_service_extension.exceptions import PksServerError
-from container_service_extension.exceptions import UnauthorizedActionError
 from container_service_extension.logger import SERVER_LOGGER as LOGGER
 from container_service_extension.ovdc_cache import OvdcCache
 from container_service_extension.pks_cache import PKS_CLUSTER_DOMAIN_KEY
@@ -23,12 +20,8 @@ from container_service_extension.server_constants import CseOperation
 from container_service_extension.server_constants import K8S_PROVIDER_KEY
 from container_service_extension.server_constants import K8sProviders
 from container_service_extension.utils import connect_vcd_user_via_token
-from container_service_extension.utils \
-    import create_pks_compute_profile_name_from_vdc_id
 from container_service_extension.utils import exception_handler
 from container_service_extension.utils import get_pks_cache
-from container_service_extension.utils import get_server_runtime_config
-from container_service_extension.utils import get_vcd_sys_admin_client
 from container_service_extension.vcdbroker import VcdBroker
 
 
@@ -56,11 +49,10 @@ class BrokerManager(object):
         self.tenant_auth_token = tenant_auth_token
         self.req_spec = request_spec
         self.pks_cache = get_pks_cache()
-        self.ovdc_cache = OvdcCache(get_vcd_sys_admin_client())
+        self.ovdc_cache = OvdcCache()
         self.is_ovdc_present_in_request = False
-        config = get_server_runtime_config()
         self.vcd_client, self.session = connect_vcd_user_via_token(
-            vcd_uri=config['vcd']['host'], tenant_auth_token=tenant_auth_token)
+            tenant_auth_token=tenant_auth_token)
 
     @exception_handler
     def invoke(self, op):
@@ -148,78 +140,55 @@ class BrokerManager(object):
                 {'cluster_name': self.req_spec.get('cluster_name'),
                  'node_name': self.req_spec.get('node_name')}
             result['body'] = self._get_node_info(**node_spec)[0]
-        elif op == CseOperation.OVDC_ENABLE_DISABLE:
-            pks_ctx, ovdc = self._get_ovdc_params()
-            if self.req_spec[K8S_PROVIDER_KEY] == K8sProviders.PKS:
-                self._create_pks_compute_profile(pks_ctx)
-            task = self.ovdc_cache. \
-                set_ovdc_container_provider_metadata(
-                    ovdc,
-                    container_prov_data=pks_ctx,
-                    container_provider=self.req_spec[K8S_PROVIDER_KEY])
-            # TODO() Constructing response should be moved out of this layer
-            result['body'] = {'task_href': task.get('href')}
-            result['status_code'] = requests.codes.accepted
-        elif op == CseOperation.OVDC_INFO:
-            ovdc_id = self.req_spec.get('ovdc_id')
-            # TODO() Constructing response should be moved out of this layer
-            result['body'] = self.ovdc_cache. \
-                get_ovdc_container_provider_metadata(ovdc_id=ovdc_id)
-        elif op == CseOperation.OVDC_LIST:
-            list_pks_plans = self.req_spec.get('list_pks_plans', False)
-            result['body'] = self._list_ovdcs(list_pks_plans=list_pks_plans)
 
         return result
 
-    def _list_ovdcs(self, list_pks_plans=False):
-        """Get list of ovdcs.
+    def _get_cluster_config(self, **cluster_spec):
+        """Get the cluster configuration.
 
-        If client is sysadmin,
-            Gets all ovdcs of all organizations.
-        Else
-            Gets all ovdcs of the organization in context.
+        :param str cluster_name: Name of cluster.
+
+        :return: Cluster config.
+
+        :rtype: str
         """
-        if self.vcd_client.is_sysadmin():
-            org_resource_list = self.vcd_client.get_org_list()
+        cluster_name = cluster_spec['cluster_name']
+        if self.is_ovdc_present_in_request:
+            broker = self.get_broker_based_on_vdc()
+            return broker.get_cluster_config(cluster_name=cluster_name)
         else:
-            org_resource_list = list(self.vcd_client.get_org())
+            cluster, broker = self._find_cluster_in_org(cluster_name)
+            if cluster:
+                return broker.get_cluster_config(cluster_name=cluster['name'])
 
-        ovdc_list = []
-        vc_to_pks_plans_map = {}
-        if list_pks_plans:
-            if self.vcd_client.is_sysadmin():
-                vc_to_pks_plans_map = self._construct_vc_to_pks_map()
-            else:
-                raise UnauthorizedActionError(
-                    'Operation Denied. Plans available only for '
-                    'System Administrator.')
-        for org_resource in org_resource_list:
-            org = Org(self.vcd_client, resource=org_resource)
-            vdc_list = org.list_vdcs()
-            for vdc in vdc_list:
-                ctr_prov_ctx = \
-                    self.ovdc_cache.get_ovdc_container_provider_metadata(
-                        ovdc_name=vdc['name'], org_name=org.get_name(),
-                        credentials_required=False)
-                if list_pks_plans:
-                    pks_plans, pks_server = self.\
-                        _get_pks_plans_and_server_for_vdc(vdc,
-                                                          org_resource,
-                                                          vc_to_pks_plans_map)
-                    vdc_dict = {
-                        'org': org.get_name(),
-                        'name': vdc['name'],
-                        'pks_api_server': pks_server,
-                        'available pks plans': pks_plans
-                    }
-                else:
-                    vdc_dict = {
-                        'name': vdc['name'],
-                        'org': org.get_name(),
-                        K8S_PROVIDER_KEY: ctr_prov_ctx[K8S_PROVIDER_KEY]
-                    }
-                ovdc_list.append(vdc_dict)
-        return ovdc_list
+        raise ClusterNotFoundError(f"Cluster {cluster_name} not found "
+                                   f"either in vCD or PKS")
+
+    def _create_cluster(self, **cluster_spec):
+        cluster_name = cluster_spec['cluster_name']
+        # 'is_org_admin_search' is used here to prevent cluster creation with
+        # same cluster-name by users within org.
+        # If it is true, cluster list is filtered by the org name of the
+        # logged-in user to check for duplicates.
+        cluster, _ = self._find_cluster_in_org(cluster_name,
+                                               is_org_admin_search=True)
+        if not cluster:
+            ctr_prov_ctx = self._get_ctr_prov_ctx_from_ovdc_metadata()
+            if ctr_prov_ctx.get(
+                    K8S_PROVIDER_KEY) == K8sProviders.PKS:
+                cluster_spec['pks_plan'] = ctr_prov_ctx[PKS_PLANS_KEY][0]
+                cluster_spec['pks_ext_host'] = f"{cluster_name}." \
+                    f"{ctr_prov_ctx[PKS_CLUSTER_DOMAIN_KEY]}"
+            broker = self._get_broker_based_on_ctr_prov_ctx(ctr_prov_ctx)
+            return broker.create_cluster(**cluster_spec)
+        else:
+            raise ClusterAlreadyExistsError(
+                f"Cluster {cluster_name} already exists.")
+
+    def _delete_cluster(self, **cluster_spec):
+        cluster_name = cluster_spec['cluster_name']
+        _, broker = self._get_cluster_info(**cluster_spec)
+        return broker.delete_cluster(cluster_name=cluster_name)
 
     def _get_cluster_info(self, **cluster_spec):
         """Get cluster details directly from cloud provider.
@@ -245,60 +214,6 @@ class BrokerManager(object):
             cluster, broker = self._find_cluster_in_org(cluster_name)
             if cluster:
                 return cluster, broker
-
-        raise ClusterNotFoundError(f"Cluster {cluster_name} not found "
-                                   f"either in vCD or PKS")
-
-    def _get_node_info(self, **node_spec):
-        """Get node details directly from cloud provider.
-
-        Logic of the method is as follows.
-
-        If 'ovdc' is present in the cluster spec,
-            choose the right broker (by identifying the container_provider
-            (vcd|pks) defined for that ovdc) to do get_node operation.
-        else
-            Invoke set of all (vCD/PKS) brokers in the org to find the cluster
-            and then do get_node operation
-
-        :return: a tuple of node information as dictionary and the broker
-            instance used to find the cluster information.
-
-        :rtype: tuple
-        """
-        cluster_name = node_spec['cluster_name']
-        node_name = node_spec['node_name']
-        if self.is_ovdc_present_in_request:
-            broker = self.get_broker_based_on_vdc()
-            return broker.get_node_info(cluster_name, node_name), broker
-        else:
-            vcd_broker = VcdBroker(self.tenant_auth_token, self.req_spec)
-            cluster = vcd_broker.get_cluster_info(cluster_name)
-            if cluster:
-                return vcd_broker.get_node_info(
-                    cluster_name, node_name), vcd_broker
-
-        raise ClusterNotFoundError(f"Cluster {cluster_name} with "
-                                   f"Node {node_name} not found "
-                                   f"either in vCD or PKS")
-
-    def _get_cluster_config(self, **cluster_spec):
-        """Get the cluster configuration.
-
-        :param str cluster_name: Name of cluster.
-
-        :return: Cluster config.
-
-        :rtype: str
-        """
-        cluster_name = cluster_spec['cluster_name']
-        if self.is_ovdc_present_in_request:
-            broker = self.get_broker_based_on_vdc()
-            return broker.get_cluster_config(cluster_name=cluster_name)
-        else:
-            cluster, broker = self._find_cluster_in_org(cluster_name)
-            if cluster:
-                return broker.get_cluster_config(cluster_name=cluster['name'])
 
         raise ClusterNotFoundError(f"Cluster {cluster_name} not found "
                                    f"either in vCD or PKS")
@@ -346,31 +261,38 @@ class BrokerManager(object):
         cluster, broker = self._get_cluster_info(**cluster_spec)
         return broker.resize_cluster(curr_cluster_info=cluster, **cluster_spec)
 
-    def _delete_cluster(self, **cluster_spec):
-        cluster_name = cluster_spec['cluster_name']
-        _, broker = self._get_cluster_info(**cluster_spec)
-        return broker.delete_cluster(cluster_name=cluster_name)
+    def _get_node_info(self, **node_spec):
+        """Get node details directly from cloud provider.
 
-    def _create_cluster(self, **cluster_spec):
-        cluster_name = cluster_spec['cluster_name']
-        # 'is_org_admin_search' is used here to prevent cluster creation with
-        # same cluster-name by users within org.
-        # If it is true, cluster list is filtered by the org name of the
-        # logged-in user to check for duplicates.
-        cluster, _ = self._find_cluster_in_org(cluster_name,
-                                               is_org_admin_search=True)
-        if not cluster:
-            ctr_prov_ctx = self._get_ctr_prov_ctx_from_ovdc_metadata()
-            if ctr_prov_ctx.get(
-                    K8S_PROVIDER_KEY) == K8sProviders.PKS:
-                cluster_spec['pks_plan'] = ctr_prov_ctx[PKS_PLANS_KEY][0]
-                cluster_spec['pks_ext_host'] = f"{cluster_name}." \
-                    f"{ctr_prov_ctx[PKS_CLUSTER_DOMAIN_KEY]}"
-            broker = self._get_broker_based_on_ctr_prov_ctx(ctr_prov_ctx)
-            return broker.create_cluster(**cluster_spec)
+        Logic of the method is as follows.
+
+        If 'ovdc' is present in the cluster spec,
+            choose the right broker (by identifying the container_provider
+            (vcd|pks) defined for that ovdc) to do get_node operation.
+        else
+            Invoke set of all (vCD/PKS) brokers in the org to find the cluster
+            and then do get_node operation
+
+        :return: a tuple of node information as dictionary and the broker
+            instance used to find the cluster information.
+
+        :rtype: tuple
+        """
+        cluster_name = node_spec['cluster_name']
+        node_name = node_spec['node_name']
+        if self.is_ovdc_present_in_request:
+            broker = self.get_broker_based_on_vdc()
+            return broker.get_node_info(cluster_name, node_name), broker
         else:
-            raise ClusterAlreadyExistsError(
-                f"Cluster {cluster_name} already exists.")
+            vcd_broker = VcdBroker(self.tenant_auth_token, self.req_spec)
+            cluster = vcd_broker.get_cluster_info(cluster_name)
+            if cluster:
+                return vcd_broker.get_node_info(
+                    cluster_name, node_name), vcd_broker
+
+        raise ClusterNotFoundError(f"Cluster {cluster_name} with "
+                                   f"Node {node_name} not found "
+                                   f"either in vCD or PKS")
 
     def _find_cluster_in_org(self, cluster_name, is_org_admin_search=False):
         """Invoke set of all (vCD/PKS)brokers in the org to find the cluster.
@@ -519,76 +441,6 @@ class BrokerManager(object):
 
         return self._get_broker_based_on_ctr_prov_ctx(ctr_prov_ctx)
 
-    def _get_ovdc_params(self):
-        ovdc_id = self.req_spec.get('ovdc_id')
-        org_name = self.req_spec.get('org_name')
-        pks_plans = self.req_spec['pks_plans']
-        pks_cluster_domain = self.req_spec['pks_cluster_domain']
-        ovdc = self.ovdc_cache.get_ovdc(ovdc_id=ovdc_id)
-        pvdc_id = self.ovdc_cache.get_pvdc_id(ovdc)
-
-        pks_context = None
-        if self.req_spec[K8S_PROVIDER_KEY] == K8sProviders.PKS:
-            if not self.pks_cache:
-                raise CseServerError('PKS config file does not exist')
-            pvdc_info = self.pks_cache.get_pvdc_info(pvdc_id)
-            if not pvdc_info:
-                LOGGER.debug(f"pvdc '{pvdc_id}' is not backed "
-                             f"by PKS-managed-vSphere resources")
-                raise CseServerError(f"'{ovdc.resource.get('name')}' is not "
-                                     f"eligible to provide resources for "
-                                     f"PKS clusters. Refer debug logs for more"
-                                     f" details.")
-            pks_account_info = self.pks_cache.get_pks_account_info(
-                org_name, pvdc_info.vc)
-            nsxt_info = self.pks_cache.get_nsxt_info(pvdc_info.vc)
-
-            pks_compute_profile_name = \
-                create_pks_compute_profile_name_from_vdc_id(ovdc_id)
-            pks_context = OvdcCache.construct_pks_context(
-                pks_account_info=pks_account_info,
-                pvdc_info=pvdc_info,
-                nsxt_info=nsxt_info,
-                pks_compute_profile_name=pks_compute_profile_name,
-                pks_plans=pks_plans,
-                pks_cluster_domain=pks_cluster_domain,
-                credentials_required=True)
-
-        return pks_context, ovdc
-
-    def _create_pks_compute_profile(self, pks_ctx):
-        ovdc_id = self.req_spec.get('ovdc_id')
-        org_name = self.req_spec.get('org_name')
-        ovdc_name = self.req_spec.get('ovdc_name')
-        # Compute profile creation
-        pks_compute_profile_name = \
-            create_pks_compute_profile_name_from_vdc_id(ovdc_id)
-        pks_compute_profile_description = f"{org_name}--{ovdc_name}" \
-            f"--{ovdc_id}"
-        pks_az_name = f"az-{ovdc_name}"
-        ovdc_rp_name = f"{ovdc_name} ({ovdc_id})"
-
-        compute_profile_params = PksComputeProfileParams(
-            pks_compute_profile_name, pks_az_name,
-            pks_compute_profile_description,
-            pks_ctx.get('cpi'),
-            pks_ctx.get('datacenter'),
-            pks_ctx.get('cluster'),
-            ovdc_rp_name).to_dict()
-
-        LOGGER.debug(f"Creating PKS Compute Profile with name:"
-                     f"{pks_compute_profile_name}")
-
-        pksbroker = PKSBroker(self.tenant_auth_token, self.req_spec, pks_ctx)
-        try:
-            pksbroker.create_compute_profile(**compute_profile_params)
-        except PksServerError as ex:
-            if ex.status == requests.codes.conflict:
-                LOGGER.debug(f"Compute profile name {pks_compute_profile_name}"
-                             f" already exists\n{str(ex)}")
-            else:
-                raise ex
-
     def _get_pks_plans_and_server_for_vdc(self,
                                           vdc,
                                           org_resource,
@@ -618,21 +470,3 @@ class BrokerManager(object):
             plan_names = [plan.get('name') for plan in plans]
             pks_vc_plans_map[pks_ctx['vc']] = [plan_names, pks_ctx['host']]
         return pks_vc_plans_map
-
-
-class PksComputeProfileParams(namedtuple("PksComputeProfileParams",
-                                         'cp_name, az_name, description,'
-                                         'cpi,datacenter_name, '
-                                         'cluster_name, ovdc_rp_name')):
-    """Construct PKS ComputeProfile Parameters ."""
-
-    def __str__(self):
-        return f"class:{PksComputeProfileParams.__name__}," \
-            f" cp_name:{self.cp_name}, az_name:{self.az_name}, " \
-            f" description:{self.description}, cpi:{self.cpi}, " \
-            f" datacenter_name:{self.datacenter_name}, " \
-            f" cluster_name:{self.cluster_name}, " \
-            f" ovdc_rp_name:{self.ovdc_rp_name}"
-
-    def to_dict(self):
-        return dict(self._asdict())
