+++ container_service_extension/cse.py
@@ -250,8 +250,7 @@ def check(ctx, config, check_install, template):
     required=False,
     default=None,
     type=click.File('r'),
-    help='SSH public key to connect to the guest OS on the VM.'
-)
+    help='SSH public key to connect to the guest OS on the VM.')
 def install(ctx, config, template, update, no_capture, ssh_key_file):
     """Install CSE on vCloud Director."""
     try:
