+++ container_service_extension/vcdbroker.py
@@ -68,16 +68,7 @@ MAX_HOST_NAME_LENGTH = 25
 ROLLBACK_FLAG = 'disable_rollback'
 
 
-def spinning_cursor():
-    while True:
-        for cursor in '|/-\\':
-            yield cursor
-
-
-spinner = spinning_cursor()
-
-
-def rollback(func):
+def supports_rollback(func):
     """Decorate to rollback on cluster and node creation failures.
 
     :param func: reference to the original function that is decorated
@@ -109,19 +100,6 @@ def rollback(func):
     return wrapper
 
 
-def task_callback(task):
-    message = f"\x1b[2K\r{task.get('operationName')}: " \
-              f"{task.get('operation')}, status: {task.get('status')}"
-    if hasattr(task, 'Progress'):
-        message += f", progress: {task.Progress}%"
-    if task.get('status').lower() in [
-            TaskStatus.QUEUED.value, TaskStatus.PENDING.value,
-            TaskStatus.PRE_RUNNING.value, TaskStatus.RUNNING.value
-    ]:
-        message += f" {spinner.next()} "
-    click.secho(message)
-
-
 class VcdBroker(AbstractBroker, threading.Thread):
     def __init__(self, tenant_auth_token, request_spec):
         super().__init__(tenant_auth_token, request_spec)
@@ -141,13 +119,15 @@ class VcdBroker(AbstractBroker, threading.Thread):
         self.cluster_id = None
         self.daemon = False
 
-    def _to_message(self, e):
-        if hasattr(e, 'message'):
-            return {'message': e.message}
-        else:
-            return {'message': str(e)}
+    def _connect_sys_admin(self):
+        self.sys_admin_client = get_vcd_sys_admin_client()
 
-    def update_task(self,
+    def _disconnect_sys_admin(self):
+        if self.sys_admin_client is not None:
+            self.sys_admin_client.logout()
+            self.sys_admin_client = None
+
+    def _update_task(self,
                     status,
                     message=None,
                     error_message=None,
@@ -184,7 +164,7 @@ class VcdBroker(AbstractBroker, threading.Thread):
             stack_trace=stack_trace
         )
 
-    def is_valid_name(self, name):
+    def _is_valid_name(self, name):
         """Validate that the cluster name against the pattern."""
         if len(name) > MAX_HOST_NAME_LENGTH:
             return False
@@ -193,7 +173,7 @@ class VcdBroker(AbstractBroker, threading.Thread):
         allowed = re.compile(r"(?!-)[A-Z\d-]{1,63}(?<!-)$", re.IGNORECASE)
         return all(allowed.match(x) for x in name.split("."))
 
-    def get_template(self, name=None):
+    def _get_template(self, name=None):
         server_config = get_server_runtime_config()
         if name is None:
             if 'template' in self.req_spec and \
@@ -205,6 +185,74 @@ class VcdBroker(AbstractBroker, threading.Thread):
             if template['name'] == name:
                 return template
         raise Exception(f"Template {name} not found.")
+
+    def _get_nfs_exports(self, ip, vapp, node):
+        """Get the exports from remote NFS server (helper method).
+
+        :param ip: (str): IP address of the NFS server
+        :param vapp: (pyvcloud.vcd.vapp.VApp): The vApp or cluster
+         to which node belongs
+        :param node: (str): IP address of the NFS server
+        :param node: (`lxml.objectify.StringElement`) object
+        representing the vm resource.
+
+        :return: (List): List of exports
+        """
+        # TODO(right template) find a right way to retrieve
+        # the template from which nfs node was created.
+        server_config = get_server_runtime_config()
+        template = server_config['broker']['templates'][0]
+        script = f"#!/usr/bin/env bash\nshowmount -e {ip}"
+        result = execute_script_in_nodes(
+            server_config, vapp, template['admin_password'],
+            script, nodes=[node], check_tools=False)
+        lines = result[0][1].content.decode().split('\n')
+        exports = []
+        for index in range(1, len(lines) - 1):
+            export = lines[index].strip().split()[0]
+            exports.append(export)
+        return exports
+
+    def node_rollback(self, node_list):
+        """Rollback for node creation failure.
+
+        :param list node_list: faulty nodes to be deleted
+        """
+        LOGGER.info(f"About to rollback nodes from cluster with name: "
+                    "{self.cluster_name}")
+        LOGGER.info(f"Node list to be deleted:{node_list}")
+        vapp = VApp(self.tenant_client, href=self.cluster['vapp_href'])
+        template = self._get_template()
+        try:
+            server_config = get_server_runtime_config()
+            delete_nodes_from_cluster(server_config, vapp, template,
+                                      node_list, force=True)
+        except Exception:
+            LOGGER.warning("Couldn't delete node {node_list} from cluster:"
+                           "{self.cluster_name}")
+        for vm_name in node_list:
+            vm = VM(self.tenant_client, resource=vapp.get_vm(vm_name))
+            try:
+                vm.undeploy()
+            except Exception:
+                LOGGER.warning(f"Couldn't undeploy VM {vm_name}")
+        vapp.delete_vms(node_list)
+        LOGGER.info(f"Successfully deleted nodes: {node_list}")
+
+    def cluster_rollback(self):
+        """Rollback for cluster creation failure."""
+        LOGGER.info(f"About to rollback cluster with name: "
+                    "{self.cluster_name}")
+        self._connect_tenant()
+        clusters = load_from_metadata(
+            self.tenant_client, name=self.cluster_name)
+        if len(clusters) != 1:
+            LOGGER.debug(f"Cluster {self.cluster_name} not found.")
+            return
+        self.cluster = clusters[0]
+        vdc = VDC(self.tenant_client, href=self.cluster['vdc_href'])
+        vdc.delete_vapp(self.cluster['name'], force=True)
+        LOGGER.info(f"Successfully deleted cluster: {self.cluster_name}")
 
     def run(self):
         LOGGER.debug(f"Thread started for operation={self.op}")
@@ -327,32 +375,21 @@ class VcdBroker(AbstractBroker, threading.Thread):
                                  f"'{cluster_name}'")
         return node_info
 
-    def _get_nfs_exports(self, ip, vapp, node):
-        """Get the exports from remote NFS server (helper method).
-
-        :param ip: (str): IP address of the NFS server
-        :param vapp: (pyvcloud.vcd.vapp.VApp): The vApp or cluster
-         to which node belongs
-        :param node: (str): IP address of the NFS server
-        :param node: (`lxml.objectify.StringElement`) object
-        representing the vm resource.
-
-        :return: (List): List of exports
-        """
-        # TODO(right template) find a right way to retrieve
-        # the template from which nfs node was created.
+    def get_cluster_config(self, cluster_name):
+        self._connect_tenant()
+        clusters = load_from_metadata(
+            self.tenant_client,
+            name=cluster_name,
+            org_name=self.req_spec.get('org'),
+            vdc_name=self.req_spec.get('vdc'))
+        if len(clusters) != 1:
+            raise CseServerError(f"Cluster '{cluster_name}' not found")
+        vapp = VApp(self.tenant_client, href=clusters[0]['vapp_href'])
+        template = self._get_template(name=clusters[0]['template'])
         server_config = get_server_runtime_config()
-        template = server_config['broker']['templates'][0]
-        script = f"#!/usr/bin/env bash\nshowmount -e {ip}"
-        result = execute_script_in_nodes(
-            server_config, vapp, template['admin_password'],
-            script, nodes=[node], check_tools=False)
-        lines = result[0][1].content.decode().split('\n')
-        exports = []
-        for index in range(1, len(lines) - 1):
-            export = lines[index].strip().split()[0]
-            exports.append(export)
-        return exports
+        result = get_cluster_config(server_config, vapp,
+                                    template['admin_password'])
+        return result
 
     @secure(required_rights=[CSE_NATIVE_DEPLOY_RIGHT_NAME])
     def create_cluster(self, cluster_name, vdc_name, node_count,
@@ -373,14 +410,14 @@ class VcdBroker(AbstractBroker, threading.Thread):
         LOGGER.debug(f"About to create cluster {cluster_name} on {vdc_name} "
                      f"with {node_count} nodes, sp={storage_profile}")
 
-        if not self.is_valid_name(cluster_name):
+        if not self._is_valid_name(cluster_name):
             raise CseServerError(f"Invalid cluster name '{cluster_name}'")
         self._connect_tenant()
         self._connect_sys_admin()
         self.cluster_name = cluster_name
         self.cluster_id = str(uuid.uuid4())
         self.op = OP_CREATE_CLUSTER
-        self.update_task(
+        self._update_task(
             TaskStatus.RUNNING,
             message=f"Creating cluster {cluster_name}({self.cluster_id})")
         self.daemon = True
@@ -391,7 +428,7 @@ class VcdBroker(AbstractBroker, threading.Thread):
         result['task_href'] = self.task_resource.get('href')
         return result
 
-    @rollback
+    @supports_rollback
     def create_cluster_thread(self):
         network_name = self.req_spec['network']
         try:
@@ -406,8 +443,8 @@ class VcdBroker(AbstractBroker, threading.Thread):
             org = Org(self.tenant_client, resource=org_resource)
             vdc_resource = org.get_vdc(self.req_spec['vdc'])
             vdc = VDC(self.tenant_client, resource=vdc_resource)
-            template = self.get_template()
-            self.update_task(
+            template = self._get_template()
+            self._update_task(
                 TaskStatus.RUNNING,
                 message=f"Creating cluster vApp {self.cluster_name}"
                         f"({self.cluster_id})")
@@ -432,7 +469,7 @@ class VcdBroker(AbstractBroker, threading.Thread):
             for k, v in tags.items():
                 task = vapp.set_metadata('GENERAL', 'READWRITE', k, v)
                 self.tenant_client.get_task_monitor().wait_for_status(task)
-            self.update_task(
+            self._update_task(
                 TaskStatus.RUNNING,
                 message=f"Creating master node for {self.cluster_name}"
                         f"({self.cluster_id})")
@@ -446,7 +483,7 @@ class VcdBroker(AbstractBroker, threading.Thread):
                 raise MasterNodeCreationError(
                     "Error while adding master node:", str(e))
 
-            self.update_task(
+            self._update_task(
                 TaskStatus.RUNNING,
                 message=f"Initializing cluster {self.cluster_name}"
                         f"({self.cluster_id})")
@@ -457,7 +494,7 @@ class VcdBroker(AbstractBroker, threading.Thread):
                                      master_ip)
             self.tenant_client.get_task_monitor().wait_for_status(task)
             if self.req_spec['node_count'] > 0:
-                self.update_task(
+                self._update_task(
                     TaskStatus.RUNNING,
                     message=f"Creating {self.req_spec['node_count']} node(s) "
                             f"for {self.cluster_name}({self.cluster_id})")
@@ -469,14 +506,14 @@ class VcdBroker(AbstractBroker, threading.Thread):
                     raise WorkerNodeCreationError(
                         "Error while creating worker node:", str(e))
 
-                self.update_task(
+                self._update_task(
                     TaskStatus.RUNNING,
                     message=f"Adding {self.req_spec['node_count']} node(s) to "
                             f"{self.cluster_name}({self.cluster_id})")
                 vapp.reload()
                 join_cluster(server_config, vapp, template)
             if self.req_spec['enable_nfs']:
-                self.update_task(
+                self._update_task(
                     TaskStatus.RUNNING,
                     message=f"Creating NFS node for {self.cluster_name}"
                             f"({self.cluster_id})")
@@ -488,7 +525,7 @@ class VcdBroker(AbstractBroker, threading.Thread):
                     raise NFSNodeCreationError(
                         "Error while creating NFS node:", str(e))
 
-            self.update_task(
+            self._update_task(
                 TaskStatus.SUCCESS,
                 message=f"Created cluster {self.cluster_name}"
                         f"({self.cluster_id})")
@@ -498,7 +535,7 @@ class VcdBroker(AbstractBroker, threading.Thread):
             LOGGER.error(traceback.format_exc())
             error_obj = error_to_json(e)
             stack_trace = ''.join(error_obj[ERROR_MESSAGE][ERROR_STACKTRACE])
-            self.update_task(
+            self._update_task(
                 TaskStatus.ERROR,
                 error_message=error_obj[ERROR_MESSAGE][ERROR_DESCRIPTION],
                 stack_trace=stack_trace)
@@ -507,7 +544,7 @@ class VcdBroker(AbstractBroker, threading.Thread):
             LOGGER.error(traceback.format_exc())
             error_obj = error_to_json(e)
             stack_trace = ''.join(error_obj[ERROR_MESSAGE][ERROR_STACKTRACE])
-            self.update_task(
+            self._update_task(
                 TaskStatus.ERROR,
                 error_message=error_obj[ERROR_MESSAGE][ERROR_DESCRIPTION],
                 stack_trace=stack_trace)
@@ -533,7 +570,7 @@ class VcdBroker(AbstractBroker, threading.Thread):
             raise CseServerError(f"Cluster {self.cluster_name} not found.")
         self.cluster = clusters[0]
         self.cluster_id = self.cluster['cluster_id']
-        self.update_task(
+        self._update_task(
             TaskStatus.RUNNING,
             message=f"Deleting cluster {self.cluster_name}"
                     f"({self.cluster_id})")
@@ -550,31 +587,15 @@ class VcdBroker(AbstractBroker, threading.Thread):
             vdc = VDC(self.tenant_client, href=self.cluster['vdc_href'])
             task = vdc.delete_vapp(self.cluster['name'], force=True)
             self.tenant_client.get_task_monitor().wait_for_status(task)
-            self.update_task(
+            self._update_task(
                 TaskStatus.SUCCESS,
                 message=f"Deleted cluster {self.cluster_name}"
                         f"({self.cluster_id})")
         except Exception as e:
             LOGGER.error(traceback.format_exc())
-            self.update_task(TaskStatus.ERROR, error_message=str(e))
+            self._update_task(TaskStatus.ERROR, error_message=str(e))
         finally:
             self._disconnect_sys_admin()
-
-    def get_cluster_config(self, cluster_name):
-        self._connect_tenant()
-        clusters = load_from_metadata(
-            self.tenant_client,
-            name=cluster_name,
-            org_name=self.req_spec.get('org'),
-            vdc_name=self.req_spec.get('vdc'))
-        if len(clusters) != 1:
-            raise CseServerError(f"Cluster '{cluster_name}' not found")
-        vapp = VApp(self.tenant_client, href=clusters[0]['vapp_href'])
-        template = self.get_template(name=clusters[0]['template'])
-        server_config = get_server_runtime_config()
-        result = get_cluster_config(server_config, vapp,
-                                    template['admin_password'])
-        return result
 
     @secure(required_rights=[CSE_NATIVE_DEPLOY_RIGHT_NAME])
     def create_nodes(self):
@@ -603,7 +624,7 @@ class VcdBroker(AbstractBroker, threading.Thread):
         self.cluster = clusters[0]
         self.op = OP_CREATE_NODES
         self.cluster_id = self.cluster['cluster_id']
-        self.update_task(
+        self._update_task(
             TaskStatus.RUNNING,
             message=f"Adding {self.req_spec['node_count']} node(s) to cluster "
                     "{self.cluster_name}({self.cluster_id})")
@@ -614,7 +635,7 @@ class VcdBroker(AbstractBroker, threading.Thread):
         result['task_href'] = self.task_resource.get('href')
         return result
 
-    @rollback
+    @supports_rollback
     def create_nodes_thread(self):
         LOGGER.debug(f"About to add nodes to cluster with name: "
                      f"{self.cluster_name}")
@@ -624,8 +645,8 @@ class VcdBroker(AbstractBroker, threading.Thread):
             org = Org(self.tenant_client, resource=org_resource)
             vdc = VDC(self.tenant_client, href=self.cluster['vdc_href'])
             vapp = VApp(self.tenant_client, href=self.cluster['vapp_href'])
-            template = self.get_template()
-            self.update_task(
+            template = self._get_template()
+            self._update_task(
                 TaskStatus.RUNNING,
                 message=f"Creating {self.req_spec['node_count']} node(s) for "
                         f"{self.cluster_name}({self.cluster_id})")
@@ -634,12 +655,12 @@ class VcdBroker(AbstractBroker, threading.Thread):
                                   server_config, self.tenant_client,
                                   org, vdc, vapp, self.req_spec)
             if self.req_spec['node_type'] == TYPE_NFS:
-                self.update_task(
+                self._update_task(
                     TaskStatus.SUCCESS,
                     message=f"Created {self.req_spec['node_count']} node(s) "
                             f"for {self.cluster_name}({self.cluster_id})")
             elif self.req_spec['node_type'] == TYPE_NODE:
-                self.update_task(
+                self._update_task(
                     TaskStatus.RUNNING,
                     message=f"Adding {self.req_spec['node_count']} node(s) to "
                             f"cluster {self.cluster_name}({self.cluster_id})")
@@ -648,7 +669,7 @@ class VcdBroker(AbstractBroker, threading.Thread):
                     target_nodes.append(spec['target_vm_name'])
                 vapp.reload()
                 join_cluster(server_config, vapp, template, target_nodes)
-                self.update_task(
+                self._update_task(
                     TaskStatus.SUCCESS,
                     message=f"Added {self.req_spec['node_count']} node(s) to "
                             f"cluster {self.cluster_name}({self.cluster_id})")
@@ -656,7 +677,7 @@ class VcdBroker(AbstractBroker, threading.Thread):
             error_obj = error_to_json(e)
             LOGGER.error(traceback.format_exc())
             stack_trace = ''.join(error_obj[ERROR_MESSAGE][ERROR_STACKTRACE])
-            self.update_task(
+            self._update_task(
                 TaskStatus.ERROR,
                 error_message=error_obj[ERROR_MESSAGE][ERROR_DESCRIPTION],
                 stack_trace=stack_trace)
@@ -665,7 +686,7 @@ class VcdBroker(AbstractBroker, threading.Thread):
             error_obj = error_to_json(e)
             LOGGER.error(traceback.format_exc())
             stack_trace = ''.join(error_obj[ERROR_MESSAGE][ERROR_STACKTRACE])
-            self.update_task(
+            self._update_task(
                 TaskStatus.ERROR,
                 error_message=error_obj[ERROR_MESSAGE][ERROR_DESCRIPTION],
                 stack_trace=stack_trace)
@@ -700,7 +721,7 @@ class VcdBroker(AbstractBroker, threading.Thread):
         self.cluster = clusters[0]
         self.op = OP_DELETE_NODES
         self.cluster_id = self.cluster['cluster_id']
-        self.update_task(
+        self._update_task(
             TaskStatus.RUNNING,
             message=f"Deleting {len(self.req_spec['nodes'])} node(s) from "
             f"cluster {self.cluster_name}({self.cluster_id})")
@@ -716,8 +737,8 @@ class VcdBroker(AbstractBroker, threading.Thread):
                      f"{self.cluster_name}")
         try:
             vapp = VApp(self.tenant_client, href=self.cluster['vapp_href'])
-            template = self.get_template()
-            self.update_task(
+            template = self._get_template()
+            self._update_task(
                 TaskStatus.RUNNING,
                 message=f"Deleting {len(self.req_spec['nodes'])} node(s) from "
                         f"{self.cluster_name}({self.cluster_id})")
@@ -731,7 +752,7 @@ class VcdBroker(AbstractBroker, threading.Thread):
             except Exception:
                 LOGGER.error(f"Couldn't delete node {self.req_spec['nodes']} "
                              f"from cluster:{self.cluster_name}")
-            self.update_task(
+            self._update_task(
                 TaskStatus.RUNNING,
                 message=f"Undeploying {len(self.req_spec['nodes'])} node(s) "
                         f"for {self.cluster_name}({self.cluster_id})")
@@ -742,13 +763,13 @@ class VcdBroker(AbstractBroker, threading.Thread):
                     self.tenant_client.get_task_monitor().wait_for_status(task)
                 except Exception:
                     LOGGER.warning(f"Couldn't undeploy VM {vm_name}")
-            self.update_task(
+            self._update_task(
                 TaskStatus.RUNNING,
                 message=f"Deleting {len(self.req_spec['nodes'])} VM(s) for "
                         f"{self.cluster_name}({self.cluster_id})")
             task = vapp.delete_vms(self.req_spec['nodes'])
             self.tenant_client.get_task_monitor().wait_for_status(task)
-            self.update_task(
+            self._update_task(
                 TaskStatus.SUCCESS,
                 message=f"Deleted {len(self.req_spec['nodes'])} node(s) to "
                         f"cluster {self.cluster_name}({self.cluster_id})")
@@ -756,53 +777,12 @@ class VcdBroker(AbstractBroker, threading.Thread):
             LOGGER.error(traceback.format_exc())
             error_obj = error_to_json(e)
             stack_trace = ''.join(error_obj[ERROR_MESSAGE][ERROR_STACKTRACE])
-            self.update_task(
+            self._update_task(
                 TaskStatus.ERROR,
                 error_message=error_obj[ERROR_MESSAGE][ERROR_DESCRIPTION],
                 stack_trace=stack_trace)
         finally:
             self._disconnect_sys_admin()
-
-    def node_rollback(self, node_list):
-        """Rollback for node creation failure.
-
-        :param list node_list: faulty nodes to be deleted
-        """
-        LOGGER.info(f"About to rollback nodes from cluster with name: "
-                    "{self.cluster_name}")
-        LOGGER.info(f"Node list to be deleted:{node_list}")
-        vapp = VApp(self.tenant_client, href=self.cluster['vapp_href'])
-        template = self.get_template()
-        try:
-            server_config = get_server_runtime_config()
-            delete_nodes_from_cluster(server_config, vapp, template,
-                                      node_list, force=True)
-        except Exception:
-            LOGGER.warning("Couldn't delete node {node_list} from cluster:"
-                           "{self.cluster_name}")
-        for vm_name in node_list:
-            vm = VM(self.tenant_client, resource=vapp.get_vm(vm_name))
-            try:
-                vm.undeploy()
-            except Exception:
-                LOGGER.warning(f"Couldn't undeploy VM {vm_name}")
-        vapp.delete_vms(node_list)
-        LOGGER.info(f"Successfully deleted nodes: {node_list}")
-
-    def cluster_rollback(self):
-        """Rollback for cluster creation failure."""
-        LOGGER.info(f"About to rollback cluster with name: "
-                    "{self.cluster_name}")
-        self._connect_tenant()
-        clusters = load_from_metadata(
-            self.tenant_client, name=self.cluster_name)
-        if len(clusters) != 1:
-            LOGGER.debug(f"Cluster {self.cluster_name} not found.")
-            return
-        self.cluster = clusters[0]
-        vdc = VDC(self.tenant_client, href=self.cluster['vdc_href'])
-        vdc.delete_vapp(self.cluster['name'], force=True)
-        LOGGER.info(f"Successfully deleted cluster: {self.cluster_name}")
 
     @secure(required_rights=[CSE_NATIVE_DEPLOY_RIGHT_NAME])
     def resize_cluster(self, cluster_name, node_count, curr_cluster_info=None):
